make get calls async

+++ WMS.Api/WMS.Services/DTOs/SaleItem/SaleItemDto.cs
@@ -3,7 +3,7 @@
 public class SaleItemDto
 {
     public int Quantity { get; init; }
-    public int UnitPrice { get; init; }
+    public decimal UnitPrice { get; init; }
     public int SaleId { get; set; }
     public int ProductId { get; init; }
     public string Product { get; init; }
